Experimental: Remove osrm profile tools
I think this library should be agnostic about osrm profiles, only
taking the path to an osrm database as config parameter.

The reasoning behind this is that users may have custom profiles.
If the option was only to choose from default car/foot/bicycle
profiles, using custom profiles would not be possible. Also, since
the data required for using this gem is probably huge, and cannot
be included in the repo, users must do some manual preparation
anyway.