Refactor code to prevent crash
The function that parses waypoint objects was vulnerable to a crash.
Root cause for this is still unknown, but a quick fix was to use
try-catch inside a loop to prevent failing items.

This is not a perfect solution, but works for now. Besides, the
code that parses results (in all OSRM services) is going to be
rewritten sooner or later, so no need to try to be perfect at this
point.